+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix update resources bug for MetadataSources and MetadataCollections([UIFC-375](https://folio-org.atlassian.net/browse/UIFC-375))
 * Fix tests and add mocks local ([UIFC-378](https://folio-org.atlassian.net/browse/UIFC-378))
 * Fix GitHub Actions workflow not running for tags ([FOLIO-4086](https://folio-org.atlassian.net/browse/FOLIO-4086))
+* Remove dependencies of k-int and stripes-erm-components ([UIFC-365](https://folio-org.atlassian.net/browse/UIFC-365))
 
 ## [6.1.0](https://github.com/folio-org/ui-finc-config/tree/v6.1.0) (2024-03-20)
 * Remove deprecated pane properties ([UIFC-337](https://issues.folio.org/browse/UIFC-337))